task/2-weather_screen-data - added api services for weather and location; added repositories ind implementations to get data from apis, added implementation to get day hour,week,current forecast;  added modules for api services, added models; cleared old code

+++ app/src/main/java/com/alenniboris/personalmanager/domain/model/weather/DayWeatherForecastModelDomain.kt
@@ -4,14 +4,14 @@ import java.util.Date
 
 data class DayWeatherForecastModelDomain(
     val dayDate: Date,
-    val temperatureMin: Int,
-    val temperatureMax: Int,
-    val feltTemperatureMin: Int,
-    val feltTemperatureMax: Int,
-    val relativeHumidity: Int,
-    val uvIndex: Int,
-    val precipitationProbability: Int,
-    val precipitation: Int,
-    val windSpeed: Int,
+    val temperatureMin: Double,
+    val temperatureMax: Double,
+    val feltTemperatureMin: Double,
+    val feltTemperatureMax: Double,
+    val relativeHumidity: Double,
+    val uvIndex: Double,
+    val precipitationProbability: Double,
+    val precipitation: Double,
+    val windSpeed: Double,
     val windDirection: WindDirection
 )
