+++ app/src/main/java/com/alenniboris/personalmanager/domain/repository/ILocationRepository.kt
@@ -6,7 +6,7 @@ import com.alenniboris.personalmanager.domain.model.common.CustomResultModelDoma
 interface ILocationRepository {
 
     suspend fun getCurrentLocation(
-        lat: Long,
-        lon: Long
+        lat: Double,
+        lon: Double
     ): CustomResultModelDomain<String, CommonExceptionModelDomain>
 }
