+++ app/src/main/java/com/alenniboris/personalmanager/domain/usecase/impl/weather_and_location/GetWeekWeatherForecastUseCaseImpl.kt
@@ -15,8 +15,8 @@ class GetWeekWeatherForecastUseCaseImpl @Inject constructor(
 ) : IGetWeekWeatherForecastUseCase {
 
     override suspend fun invoke(
-        lat: Long,
-        lon: Long
+        lat: Double,
+        lon: Double
     ): CustomResultModelDomain<List<DayWeatherForecastModelDomain>, CommonExceptionModelDomain> =
         withContext(dispatchers.IO) {
             return@withContext weatherRepository.getDayWeatherForecastForWeek(
