+++ app/src/main/java/com/alenniboris/personalmanager/domain/usecase/logic/weather_and_location/IGetWeekWeatherForecastUseCase.kt
@@ -6,7 +6,7 @@ import com.alenniboris.personalmanager.domain.model.weather.DayWeatherForecastMo
 
 interface IGetWeekWeatherForecastUseCase {
     suspend fun invoke(
-        lat: Long,
-        lon: Long
+        lat: Double,
+        lon: Double
     ): CustomResultModelDomain<List<DayWeatherForecastModelDomain>, CommonExceptionModelDomain>
 }
